changed the datatype of co-ordinates
int -> float

+++ distance_between_2_points.c
@@ -1,38 +1,38 @@
 #include<stdio.h>
 #include<math.h>
 
-int input_x_co_ordinate_of_ith_point(int ith_point)
+float input_x_co_ordinate_of_ith_point(int ith_point)
 {
-    int x;
+    float x;
     printf("Enter the x co-cordinate of the point %d: ", ith_point);
-    scanf("%d", &x);
+    scanf("%f", &x);
     return x;
 }
 
-int input_y_co_ordinate_of_ith_point(int ith_point)
+float input_y_co_ordinate_of_ith_point(int ith_point)
 {
-    int y;
+    float y;
     printf("Enter the y co-cordinate of the point %d: ", ith_point);
-    scanf("%d", &y);
+    scanf("%f", &y);
     return y;
 }
 
-float distance_formula(int x1, int y1, int x2, int y2)
+float distance_formula(float x1, float y1, float x2, float y2)
 {
     return(sqrt(pow((x2-x1), 2)+ pow((y2-y1), 2)));
 }
 
-void output(int x1, int y1, int x2, int y2, float distance)
+void output(float x1, float y1, float x2, float y2, float distance)
 {
-    printf("Distance between points (%d, %d) and (%d, %d) is %f\n", x1, y1, x2, y2, distance);
+    printf("Distance between points (%f, %f) and (%f, %f) is %f\n", x1, y1, x2, y2, distance);
 }
 
 int main()
 {
-    int x1=input_x_co_ordinate_of_ith_point(1);
-    int y1=input_y_co_ordinate_of_ith_point(1);
-    int x2=input_x_co_ordinate_of_ith_point(2);
-    int y2=input_y_co_ordinate_of_ith_point(2);
+    float x1=input_x_co_ordinate_of_ith_point(1);
+    float y1=input_y_co_ordinate_of_ith_point(1);
+    float x2=input_x_co_ordinate_of_ith_point(2);
+    float y2=input_y_co_ordinate_of_ith_point(2);
     float distance = distance_formula(x1, y1, x2,  y2);
     output(x1, y1, x2, y2, distance);
     return 0;
